Last changes before deadline
>UML diagram added
>Doxygen added
>Auto card type difener
>Minor updates

+++ OOP_Project_Online_Book_Store/CreditCard.cpp
@@ -17,14 +17,14 @@ CreditCard::~CreditCard()
 /*!
 \return Number of credit card
 */
-long CreditCard::getNumber() const
+long long CreditCard::getNumber() const
 {
 	return number;
 }
 /*!
 \param number Number of credit card will be set to this
 */
-void CreditCard::setNumber(long number)
+void CreditCard::setNumber(long long number)
 {
 	this->number = number;
 }
